requirements.txt
python-telegram-bot==13.15
google-generativeai==0.8.3
python-dotenv==1.0.0
apscheduler==3.9.1
pytz==2024.1
